Simple simulation tooling (#1)
* Simulation added

* Add first calculatable statistics

* Network simulation added

* Remove warnings

* Node entity revorked. Remove monitor notion

* simulator tests

* Continuous removal of monitors

* Working example with Network Query

* Minor refactoring

* Minor refactoring

* Add Backlog Estimation

* Backlog estimation bug

* Backlog estimation test corrected

* Update test to use breeze viz

* Update test for more reality

+++ src/main/scala/qn/distribution/Distribution.scala
@@ -119,9 +119,10 @@ case class Singular(value: Double)(implicit rand: RandBasis = Rand) extends Disc
 }
 
 case class EmpiricDistribution(values: Array[Double])(implicit rand: RandBasis = Rand) extends DiscreteDistr[Double] with Moments[Double, Double] {
+
   override def probabilityOf(x: Double): Double = if (values.size <= 0) 0 else values.count(_ < x) / values.size
 
-  override def draw(): Double = values(rand.randInt(values.size).draw())
+  override def draw(): Double = values(rand.randInt(values.size).draw()).toDouble
 
   override def mean: Double = breeze.stats.mean(values)
